Image Classification With RESNET101
In this study, we investigate the use of RESNET101 deep learning architecture for the problem of ground image classification.

A dataset of ground images from different geographical regions is used in the study. The dataset is prepared for image preprocessing and model training.

RESNET101 is a convolutional neural network (CNN) architecture that has achieved successful results in the field of image classification. In this study, the RESNET101 architecture is adapted and trained for ground image classification.

During the training process, various techniques were used to optimize the hyperparameters. The performance of the model is evaluated on test data.

The results show that the RESNET101 architecture can be effectively used for ground image classification. The proposed approach can be useful in applications in geographic information systems, remote sensing and related fields.

Future studies may focus on using different deep learning architectures and algorithms, increasing data diversity and improving model performance. Thus, more advanced and reliable systems can be created for the classification of ground images.